drop rpath patch
This rpath patch was introduced to fix glib-feedstock/issues/40.
As stated in e655eb232, this is no longer needed, it will cause meson
to skip fixing some should-be-fix RPATH on macos, making meson behaving
differently from other packaging system.

If someone does need this, it should be submitted to upstream instead of
keeping it as privite patch in conda-forge.